upgrade compile errors

+++ AlphaClient/AlphaGraph/CRenderCommandMgr.cpp
@@ -12,7 +12,7 @@ CRenderCommandMgr::~CRenderCommandMgr()
 {
 }
 
-void CRenderCommandMgr::DrawPrimitive( const SMaterial& material, EPrimitiveType primitiveType, uint16 vertexCnt, uint16 primitiveCnt, uint8 vertexFormat, uint16 vertexStride, const void* arrVertex, const void* arrIndex )
+void CRenderCommandMgr::DrawPrimitive( const SMaterial& material, EPrimitiveType primitiveType, uint16 vertexCnt, uint16 primitiveCnt, uint16 vertexFormat, uint16 vertexStride, const void* arrVertex, const void* arrIndex )
 {
 	m_sPrimiveInfo.nPrimitiveGroupCnt ++;
 	m_sPrimiveInfo.nVertexCnt += vertexCnt;
@@ -20,7 +20,7 @@ void CRenderCommandMgr::DrawPrimitive( const SMaterial& material, EPrimitiveType
 	m_pGraphic->GetRenderStateMgr().Apply(material, primitiveType, vertexCnt, primitiveCnt, vertexFormat, vertexStride, arrVertex, arrIndex);
 }
 
-void CRenderCommandMgr::DrawPrimitive( const SMaterial& material, EPrimitiveType primitiveType, uint16 vertexCnt, uint16 primitiveCnt, uint8 vertexFormat, CGeometryBuffer* vertexBuf, CGeometryBuffer* indexBuf )
+void CRenderCommandMgr::DrawPrimitive( const SMaterial& material, EPrimitiveType primitiveType, uint16 vertexCnt, uint16 primitiveCnt, uint16 vertexFormat, CGeometryBuffer* vertexBuf, CGeometryBuffer* indexBuf )
 {
 	m_sPrimiveInfo.nPrimitiveGroupCnt ++;
 	m_sPrimiveInfo.nVertexCnt += vertexCnt;
